bugfix: int overflow

+++ mars/comm/http.cc
@@ -914,15 +914,15 @@ Parser::TRecvStatus Parser::Recv(const void* _buffer, size_t _length, size_t* co
                             }
                         }
                     } else {  // no chunk
-                        int contentLength = headfields_.ContentLength();
-                        int appendlen = 0;
+                        int64_t contentLength = headfields_.ContentLength();
+                        int64_t appendlen = 0;
                         if (Fields().IsConnectionClose() && 0==contentLength) {
-                            appendlen = int(recvbuf_.Length());
-                        } else if (int(recvbuf_.Length() + bodyreceiver_->Length()) <= contentLength)
-                            appendlen = int(recvbuf_.Length());
+                            appendlen = int64_t(recvbuf_.Length());
+                        } else if (int64_t(recvbuf_.Length() + bodyreceiver_->Length()) <= contentLength)
+                            appendlen = int64_t(recvbuf_.Length());
                         else {
                             xwarn2(TSF"recv len bigger than contentlen, (%_, %_, %_)", recvbuf_.Length(), bodyreceiver_->Length(), contentLength);
-                            appendlen = contentLength - int(bodyreceiver_->Length());
+                            appendlen = contentLength - int64_t(bodyreceiver_->Length());
                         }
                         
                         bodyreceiver_->AppendData(recvbuf_.Ptr(), (size_t)appendlen);
@@ -932,7 +932,7 @@ Parser::TRecvStatus Parser::Recv(const void* _buffer, size_t _length, size_t* co
                             *consumed_bytes = origin_size - recvbuf_.Length();
                         }
                         
-                        if ((int)bodyreceiver_->Length() == contentLength) {
+                        if ((int64_t)bodyreceiver_->Length() == contentLength) {
                             recvstatus_ = kEnd;
                             bodyreceiver_->EndData();
                             return  recvstatus_;
